extract `v8pp::detail::type_info` into a header file
... for better modularization.
