+++ test/main.cpp
@@ -13,6 +13,7 @@
 
 void run_tests()
 {
+	void test_type_info();
 	void test_utility();
 	void test_context();
 	void test_convert();
@@ -29,6 +30,7 @@ void run_tests()
 
 	std::pair<char const*, void (*)()> tests[] =
 	{
+		{ "test_type_info", test_type_info },
 		{"test_utility", test_utility},
 		{"test_context", test_context},
 		{"test_convert", test_convert},
