+++ test/test_call_from_v8.cpp
@@ -2,6 +2,7 @@
 #include "v8pp/context.hpp"
 #include "v8pp/function.hpp"
 #include "v8pp/ptr_traits.hpp"
+#include "v8pp/throw_ex.hpp"
 
 #include "test.hpp"
